+++ test/test_validateDir.js
@@ -1,0 +1,11 @@
+const jsPackTools = require('../index');
+let u = new jsPackTools();
+
+describe('SUITE - validateDir()', function () {
+
+    it('TEST # 1, validate if folder exit', function () {
+        // Logic: validate if folder exist else not exist create new folder
+        u.validateDir();
+    });
+
+});