+++ CHANGELOG.md
@@ -1,6 +1,208 @@
-# Change Log
+# Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+### [0.2.1](https://github.com/jasp402/js-packtools/compare/v0.2.0...v0.2.1) (2019-05-22)
+
+
+### Bug Fixes
+
+* problem with the logic in constructor ([b0ed1df](https://github.com/jasp402/js-packtools/commit/b0ed1df))
+
+
+
+## [0.2.0](https://github.com/jasp402/js-packtools/compare/v0.1.20...v0.2.0) (2019-05-22)
+
+
+### Bug Fixes
+
+* add createFolders() in API functions ([4b663af](https://github.com/jasp402/js-packtools/commit/4b663af))
+* formatSeconds.js add jsDoc ([5f6b3f7](https://github.com/jasp402/js-packtools/commit/5f6b3f7))
+* the color parameter this bad assign ([98a6ff6](https://github.com/jasp402/js-packtools/commit/98a6ff6))
+* Update README.md ([b720c6a](https://github.com/jasp402/js-packtools/commit/b720c6a))
+
+
+### Features
+
+* changes in the logic of constructor ([fd24f72](https://github.com/jasp402/js-packtools/commit/fd24f72))
+
+
+
+## [0.1.21](https://github.com/jasp402/js-packtools/compare/v0.1.18...v0.1.21) (2019-05-21)
+
+
+### Bug Fixes
+
+* ClearFolder complete code ([013be71](https://github.com/jasp402/js-packtools/commit/013be71))
+* constant not found ([92a736e](https://github.com/jasp402/js-packtools/commit/92a736e))
+* csvToJson.js - add comment about improvement ([3578c84](https://github.com/jasp402/js-packtools/commit/3578c84))
+* deleteFile.js - add jsDoc ([25591bc](https://github.com/jasp402/js-packtools/commit/25591bc))
+* formatSeconds.js add jsDoc ([5f6b3f7](https://github.com/jasp402/js-packtools/commit/5f6b3f7))
+* test: ClearFolder() ([56f3a31](https://github.com/jasp402/js-packtools/commit/56f3a31))
+* update Readme. Include table of the content ([08a159a](https://github.com/jasp402/js-packtools/commit/08a159a))
+* Update Readme.md. fix table of content ([9832a7b](https://github.com/jasp402/js-packtools/commit/9832a7b))
+
+
+
+## [0.1.20](https://github.com/jasp402/js-packtools/compare/v0.1.19...v0.1.20) (2019-05-14)
+
+
+### Bug Fixes
+
+* csvToJson.js - add comment about improvement ([3578c84](https://github.com/jasp402/js-packtools/commit/3578c84))
+* deleteFile.js - add jsDoc ([25591bc](https://github.com/jasp402/js-packtools/commit/25591bc))
+* test: ClearFolder() ([56f3a31](https://github.com/jasp402/js-packtools/commit/56f3a31))
+* Update Readme.md. fix table of content ([9832a7b](https://github.com/jasp402/js-packtools/commit/9832a7b))
+
+
+### Tests
+
+* test_differenceDay.js ([64083bb](https://github.com/jasp402/js-packtools/commit/64083bb))
+
+
+
+### [0.1.19](https://github.com/jasp402/js-packtools/compare/v0.1.16...v0.1.19) (2019-05-13)
+
+
+### Bug Fixes
+
+* add badges of trello ([6b25965](https://github.com/jasp402/js-packtools/commit/6b25965))
+* ClearFolder complete code ([013be71](https://github.com/jasp402/js-packtools/commit/013be71))
+* constant not found ([92a736e](https://github.com/jasp402/js-packtools/commit/92a736e))
+* delete comment ([a1990ca](https://github.com/jasp402/js-packtools/commit/a1990ca))
+* jasp402/js-packtools/clearFoder.js change in documentation jsDOC ([63b0e56](https://github.com/jasp402/js-packtools/commit/63b0e56))
+* test: validateDir() ([0bddd55](https://github.com/jasp402/js-packtools/commit/0bddd55))
+* Update in test. ([58fa373](https://github.com/jasp402/js-packtools/commit/58fa373))
+* update Readme. Include table of the content ([08a159a](https://github.com/jasp402/js-packtools/commit/08a159a))
+* update README.md ([71dec4b](https://github.com/jasp402/js-packtools/commit/71dec4b))
+
+
+
+## [0.1.18](https://github.com/jasp402/js-packtools/compare/v0.1.17...v0.1.18) (2019-05-12)
+
+
+### Bug Fixes
+
+* jasp402/js-packtools/clearFoder.js change in documentation jsDOC ([63b0e56](https://github.com/jasp402/js-packtools/commit/63b0e56))
+* test: validateDir() ([0bddd55](https://github.com/jasp402/js-packtools/commit/0bddd55))
+* update README.md ([71dec4b](https://github.com/jasp402/js-packtools/commit/71dec4b))
+
+
+
+## [0.1.17](https://github.com/jasp402/js-packtools/compare/v0.1.15...v0.1.17) (2019-05-12)
+
+
+### Bug Fixes
+
+* add badges of trello ([6b25965](https://github.com/jasp402/js-packtools/commit/6b25965))
+* add test_customDate.js ([b2a9bd5](https://github.com/jasp402/js-packtools/commit/b2a9bd5))
+* ClearFolder.js - add jsDoc ([d0f1869](https://github.com/jasp402/js-packtools/commit/d0f1869))
+* createFolders.js - add jsDoc ([ebc78a0](https://github.com/jasp402/js-packtools/commit/ebc78a0))
+* csvToJson.js ([54e3e30](https://github.com/jasp402/js-packtools/commit/54e3e30))
+* delete comment ([a1990ca](https://github.com/jasp402/js-packtools/commit/a1990ca))
+* refactor: delete spaces ([1295709](https://github.com/jasp402/js-packtools/commit/1295709))
+* refactor: index.js ([e5e5575](https://github.com/jasp402/js-packtools/commit/e5e5575))
+* Update in test. ([58fa373](https://github.com/jasp402/js-packtools/commit/58fa373))
+
+
+
+## [0.1.16](https://github.com/jasp402/js-packtools/compare/v0.1.14...v0.1.16) (2019-05-09)
+
+
+### Bug Fixes
+
+* add test_customDate.js ([b2a9bd5](https://github.com/jasp402/js-packtools/commit/b2a9bd5))
+* ClearFolder.js - add jsDoc ([d0f1869](https://github.com/jasp402/js-packtools/commit/d0f1869))
+* createFolders.js - add jsDoc ([ebc78a0](https://github.com/jasp402/js-packtools/commit/ebc78a0))
+* csvToJson.js ([54e3e30](https://github.com/jasp402/js-packtools/commit/54e3e30))
+* implementation of the prototype method ([f2c17cc](https://github.com/jasp402/js-packtools/commit/f2c17cc))
+* refactor: delete spaces ([1295709](https://github.com/jasp402/js-packtools/commit/1295709))
+* refactor: index.js ([e5e5575](https://github.com/jasp402/js-packtools/commit/e5e5575))
+
+
+
+## [0.1.15](https://github.com/jasp402/js-packtools/compare/v0.1.12...v0.1.15) (2019-05-09)
+
+
+### Bug Fixes
+
+* added files to .gitignore ([adda3ce](https://github.com/jasp402/js-packtools/commit/adda3ce))
+* implementation of the prototype method ([f2c17cc](https://github.com/jasp402/js-packtools/commit/f2c17cc))
+* refactorized differenceDay() ([8ba4c35](https://github.com/jasp402/js-packtools/commit/8ba4c35))
+* Removed favicon ([2e195a3](https://github.com/jasp402/js-packtools/commit/2e195a3))
+* update .gitignore ([11569cc](https://github.com/jasp402/js-packtools/commit/11569cc))
+
+
+
+## [0.1.14](https://github.com/jasp402/js-packtools/compare/v0.1.13...v0.1.14) (2019-05-08)
+
+
+### Bug Fixes
+
+* added files to .gitignore ([adda3ce](https://github.com/jasp402/js-packtools/commit/adda3ce))
+* Removed favicon ([2e195a3](https://github.com/jasp402/js-packtools/commit/2e195a3))
+* update .gitignore ([11569cc](https://github.com/jasp402/js-packtools/commit/11569cc))
+
+
+
+### [0.1.13](https://github.com/jasp402/js-packtools/compare/v0.1.8...v0.1.13) (2019-05-08)
+
+
+### Bug Fixes
+
+* add const require ([43c83d5](https://github.com/jasp402/js-packtools/commit/43c83d5))
+* add dependency lodash@4.17.11 in package.json ([96478d4](https://github.com/jasp402/js-packtools/commit/96478d4))
+* add objectToDate ([79b21e3](https://github.com/jasp402/js-packtools/commit/79b21e3))
+* load list of function in the class and API Doc. ([277ef89](https://github.com/jasp402/js-packtools/commit/277ef89))
+* refactorized differenceDay() ([8ba4c35](https://github.com/jasp402/js-packtools/commit/8ba4c35))
+* update API Doc. ([807c1ed](https://github.com/jasp402/js-packtools/commit/807c1ed))
+* update doc/index.html ([063e26f](https://github.com/jasp402/js-packtools/commit/063e26f))
+* update Documentation ([2dba3c9](https://github.com/jasp402/js-packtools/commit/2dba3c9))
+* update README.md. ([b04a028](https://github.com/jasp402/js-packtools/commit/b04a028))
+* update test ([accbe9b](https://github.com/jasp402/js-packtools/commit/accbe9b))
+
+
+
+## [0.1.12](https://github.com/jasp402/js-packtools/compare/v0.1.11...v0.1.12) (2019-05-08)
+
+
+### Bug Fixes
+
+* add objectToDate ([79b21e3](https://github.com/jasp402/js-packtools/commit/79b21e3))
+* update test ([accbe9b](https://github.com/jasp402/js-packtools/commit/accbe9b))
+
+
+
+## [0.1.11](https://github.com/jasp402/js-packtools/compare/v0.1.10...v0.1.11) (2019-05-07)
+
+
+### Bug Fixes
+
+* add const require ([43c83d5](https://github.com/jasp402/js-packtools/commit/43c83d5))
+
+
+
+## [0.1.10](https://github.com/jasp402/js-packtools/compare/v0.1.9...v0.1.10) (2019-05-07)
+
+
+### Bug Fixes
+
+* add dependency lodash@4.17.11 in package.json ([96478d4](https://github.com/jasp402/js-packtools/commit/96478d4))
+
+
+
+## [0.1.9](https://github.com/jasp402/js-packtools/compare/v0.1.8...v0.1.9) (2019-05-07)
+
+
+### Bug Fixes
+
+* load list of function in the class and API Doc. ([277ef89](https://github.com/jasp402/js-packtools/commit/277ef89))
+* update API Doc. ([807c1ed](https://github.com/jasp402/js-packtools/commit/807c1ed))
+* update doc/index.html ([063e26f](https://github.com/jasp402/js-packtools/commit/063e26f))
+* update Documentation ([2dba3c9](https://github.com/jasp402/js-packtools/commit/2dba3c9))
+* update README.md. ([b04a028](https://github.com/jasp402/js-packtools/commit/b04a028))
+
+
 
 ## [0.1.8](https://github.com/jasp402/js-packtools/compare/v0.1.7...v0.1.8) (2019-05-06)